Doc changes for the examples.
Examples docstrings can now be extracted by sphinx.
